Login security

+++ src/main/java/com/phamhieu/bookapi/domain/book/BookService.java
@@ -1,8 +1,11 @@
 package com.phamhieu.bookapi.domain.book;
 
+import com.phamhieu.bookapi.domain.auth.AuthsProvider;
 import com.phamhieu.bookapi.persistence.book.BookStore;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import static com.phamhieu.bookapi.domain.auth.AuthError.supplyBookAccessDenied;
 import static com.phamhieu.bookapi.domain.book.BookValidation.*;
 
 import java.time.Instant;
@@ -16,6 +19,8 @@ import static com.phamhieu.bookapi.domain.book.BookError.*;
 public class BookService {
 
     private final BookStore bookStore;
+
+    private final AuthsProvider authsProvider;
 
     public List<Book> findAll() {
         return bookStore.findAll();
@@ -33,26 +38,36 @@ public class BookService {
     public Book create(final Book book) {
         validateBook(book);
 
+        book.setUserId(authsProvider.getCurrentAuthentication().getUserId());
         book.setCreatedAt(Instant.now());
         return bookStore.create(book);
     }
 
     public Book update(final UUID bookId, final Book book) {
-        validateBook(book);
+        validatePermissionWhenChangeBook(book);
 
+        validateBook(book);
         final Book existingBook = findById(bookId);
         existingBook.setTitle(book.getTitle());
         existingBook.setAuthor(book.getAuthor());
         existingBook.setDescription(book.getDescription());
         existingBook.setUpdatedAt(Instant.now());
         existingBook.setImage(book.getImage());
-        existingBook.setUserId(book.getUserId());
 
         return bookStore.update(existingBook);
     }
 
     public void delete(final UUID bookId) {
         final Book book = findById(bookId);
+        validatePermissionWhenChangeBook(book);
+
         bookStore.delete(book.getId());
+    }
+
+    private void validatePermissionWhenChangeBook(final Book book) {
+        if (authsProvider.getCurrentRole().equals("CONTRIBUTOR") &&
+                !authsProvider.getCurrentUserId().equals(book.getUserId())) {
+            throw supplyBookAccessDenied().get();
+        }
     }
 }
